gh-139322: Create test_os.test_process
