+++ Lib/test/test_os/test_posix.py
@@ -18,8 +18,8 @@ import stat
 import tempfile
 import unittest
 import warnings
-import textwrap
 from contextlib import contextmanager
+from .utils import requires_sched
 
 try:
     import posix
@@ -39,18 +39,6 @@ requires_32b = unittest.skipUnless(
     sys.maxsize < 2**32 and not (support.is_emscripten or support.is_wasi),
     'test is only meaningful on 32-bit builds'
 )
-
-def _supports_sched():
-    if not hasattr(posix, 'sched_getscheduler'):
-        return False
-    try:
-        posix.sched_getscheduler(0)
-    except OSError as e:
-        if e.errno == errno.ENOSYS:
-            return False
-    return True
-
-requires_sched = unittest.skipUnless(_supports_sched(), 'requires POSIX scheduler API')
 
 
 class PosixTester(unittest.TestCase):
@@ -1827,397 +1815,6 @@ class PosixGroupsTester(unittest.TestCase):
         for groups in [[0], list(range(16))]:
             posix.setgroups(groups)
             self.assertListEqual(groups, posix.getgroups())
-
-
-class _PosixSpawnMixin:
-    # Program which does nothing and exits with status 0 (success)
-    NOOP_PROGRAM = (sys.executable, '-I', '-S', '-c', 'pass')
-    spawn_func = None
-
-    def python_args(self, *args):
-        # Disable site module to avoid side effects. For example,
-        # on Fedora 28, if the HOME environment variable is not set,
-        # site._getuserbase() calls pwd.getpwuid() which opens
-        # /var/lib/sss/mc/passwd but then leaves the file open which makes
-        # test_close_file() to fail.
-        return (sys.executable, '-I', '-S', *args)
-
-    def test_returns_pid(self):
-        pidfile = os_helper.TESTFN
-        self.addCleanup(os_helper.unlink, pidfile)
-        script = f"""if 1:
-            import os
-            with open({pidfile!r}, "w") as pidfile:
-                pidfile.write(str(os.getpid()))
-            """
-        args = self.python_args('-c', script)
-        pid = self.spawn_func(args[0], args, os.environ)
-        support.wait_process(pid, exitcode=0)
-        with open(pidfile, encoding="utf-8") as f:
-            self.assertEqual(f.read(), str(pid))
-
-    def test_no_such_executable(self):
-        no_such_executable = 'no_such_executable'
-        try:
-            pid = self.spawn_func(no_such_executable,
-                                  [no_such_executable],
-                                  os.environ)
-        # bpo-35794: PermissionError can be raised if there are
-        # directories in the $PATH that are not accessible.
-        except (FileNotFoundError, PermissionError) as exc:
-            self.assertEqual(exc.filename, no_such_executable)
-        else:
-            pid2, status = os.waitpid(pid, 0)
-            self.assertEqual(pid2, pid)
-            self.assertNotEqual(status, 0)
-
-    def test_specify_environment(self):
-        envfile = os_helper.TESTFN
-        self.addCleanup(os_helper.unlink, envfile)
-        script = f"""if 1:
-            import os
-            with open({envfile!r}, "w", encoding="utf-8") as envfile:
-                envfile.write(os.environ['foo'])
-        """
-        args = self.python_args('-c', script)
-        pid = self.spawn_func(args[0], args,
-                              {**os.environ, 'foo': 'bar'})
-        support.wait_process(pid, exitcode=0)
-        with open(envfile, encoding="utf-8") as f:
-            self.assertEqual(f.read(), 'bar')
-
-    def test_none_file_actions(self):
-        pid = self.spawn_func(
-            self.NOOP_PROGRAM[0],
-            self.NOOP_PROGRAM,
-            os.environ,
-            file_actions=None
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_empty_file_actions(self):
-        pid = self.spawn_func(
-            self.NOOP_PROGRAM[0],
-            self.NOOP_PROGRAM,
-            os.environ,
-            file_actions=[]
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_resetids_explicit_default(self):
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', 'pass'],
-            os.environ,
-            resetids=False
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_resetids(self):
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', 'pass'],
-            os.environ,
-            resetids=True
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_setpgroup(self):
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', 'pass'],
-            os.environ,
-            setpgroup=os.getpgrp()
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_setpgroup_wrong_type(self):
-        with self.assertRaises(TypeError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setpgroup="023")
-
-    @unittest.skipUnless(hasattr(signal, 'pthread_sigmask'),
-                           'need signal.pthread_sigmask()')
-    def test_setsigmask(self):
-        code = textwrap.dedent("""\
-            import signal
-            signal.raise_signal(signal.SIGUSR1)""")
-
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', code],
-            os.environ,
-            setsigmask=[signal.SIGUSR1]
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_setsigmask_wrong_type(self):
-        with self.assertRaises(TypeError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigmask=34)
-        with self.assertRaises(TypeError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigmask=["j"])
-        with self.assertRaises(ValueError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigmask=[signal.NSIG,
-                                                    signal.NSIG+1])
-
-    def test_setsid(self):
-        rfd, wfd = os.pipe()
-        self.addCleanup(os.close, rfd)
-        try:
-            os.set_inheritable(wfd, True)
-
-            code = textwrap.dedent(f"""
-                import os
-                fd = {wfd}
-                sid = os.getsid(0)
-                os.write(fd, str(sid).encode())
-            """)
-
-            try:
-                pid = self.spawn_func(sys.executable,
-                                      [sys.executable, "-c", code],
-                                      os.environ, setsid=True)
-            except NotImplementedError as exc:
-                self.skipTest(f"setsid is not supported: {exc!r}")
-            except PermissionError as exc:
-                self.skipTest(f"setsid failed with: {exc!r}")
-        finally:
-            os.close(wfd)
-
-        support.wait_process(pid, exitcode=0)
-
-        output = os.read(rfd, 100)
-        child_sid = int(output)
-        parent_sid = os.getsid(os.getpid())
-        self.assertNotEqual(parent_sid, child_sid)
-
-    @unittest.skipUnless(hasattr(signal, 'pthread_sigmask'),
-                         'need signal.pthread_sigmask()')
-    def test_setsigdef(self):
-        original_handler = signal.signal(signal.SIGUSR1, signal.SIG_IGN)
-        code = textwrap.dedent("""\
-            import signal
-            signal.raise_signal(signal.SIGUSR1)""")
-        try:
-            pid = self.spawn_func(
-                sys.executable,
-                [sys.executable, '-c', code],
-                os.environ,
-                setsigdef=[signal.SIGUSR1]
-            )
-        finally:
-            signal.signal(signal.SIGUSR1, original_handler)
-
-        support.wait_process(pid, exitcode=-signal.SIGUSR1)
-
-    def test_setsigdef_wrong_type(self):
-        with self.assertRaises(TypeError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigdef=34)
-        with self.assertRaises(TypeError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigdef=["j"])
-        with self.assertRaises(ValueError):
-            self.spawn_func(sys.executable,
-                            [sys.executable, "-c", "pass"],
-                            os.environ, setsigdef=[signal.NSIG, signal.NSIG+1])
-
-    @requires_sched
-    @unittest.skipIf(sys.platform.startswith(('freebsd', 'netbsd')),
-                     "bpo-34685: test can fail on BSD")
-    def test_setscheduler_only_param(self):
-        policy = os.sched_getscheduler(0)
-        priority = os.sched_get_priority_min(policy)
-        code = textwrap.dedent(f"""\
-            import os, sys
-            if os.sched_getscheduler(0) != {policy}:
-                sys.exit(101)
-            if os.sched_getparam(0).sched_priority != {priority}:
-                sys.exit(102)""")
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', code],
-            os.environ,
-            scheduler=(None, os.sched_param(priority))
-        )
-        support.wait_process(pid, exitcode=0)
-
-    @requires_sched
-    @unittest.skipIf(sys.platform.startswith(('freebsd', 'netbsd')),
-                     "bpo-34685: test can fail on BSD")
-    @unittest.skipIf(platform.libc_ver()[0] == 'glibc' and
-                     os.sched_getscheduler(0) in [
-                        os.SCHED_BATCH,
-                        os.SCHED_IDLE,
-                        os.SCHED_DEADLINE],
-                     "Skip test due to glibc posix_spawn policy")
-    def test_setscheduler_with_policy(self):
-        policy = os.sched_getscheduler(0)
-        priority = os.sched_get_priority_min(policy)
-        code = textwrap.dedent(f"""\
-            import os, sys
-            if os.sched_getscheduler(0) != {policy}:
-                sys.exit(101)
-            if os.sched_getparam(0).sched_priority != {priority}:
-                sys.exit(102)""")
-        pid = self.spawn_func(
-            sys.executable,
-            [sys.executable, '-c', code],
-            os.environ,
-            scheduler=(policy, os.sched_param(priority))
-        )
-        support.wait_process(pid, exitcode=0)
-
-    def test_multiple_file_actions(self):
-        file_actions = [
-            (os.POSIX_SPAWN_OPEN, 3, os.path.realpath(__file__), os.O_RDONLY, 0),
-            (os.POSIX_SPAWN_CLOSE, 0),
-            (os.POSIX_SPAWN_DUP2, 1, 4),
-        ]
-        pid = self.spawn_func(self.NOOP_PROGRAM[0],
-                              self.NOOP_PROGRAM,
-                              os.environ,
-                              file_actions=file_actions)
-        support.wait_process(pid, exitcode=0)
-
-    def test_bad_file_actions(self):
-        args = self.NOOP_PROGRAM
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[None])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[()])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(None,)])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(12345,)])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(os.POSIX_SPAWN_CLOSE,)])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(os.POSIX_SPAWN_CLOSE, 1, 2)])
-        with self.assertRaises(TypeError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(os.POSIX_SPAWN_CLOSE, None)])
-        with self.assertRaises(ValueError):
-            self.spawn_func(args[0], args, os.environ,
-                            file_actions=[(os.POSIX_SPAWN_OPEN,
-                                           3, __file__ + '\0',
-                                           os.O_RDONLY, 0)])
-
-    def test_open_file(self):
-        outfile = os_helper.TESTFN
-        self.addCleanup(os_helper.unlink, outfile)
-        script = """if 1:
-            import sys
-            sys.stdout.write("hello")
-            """
-        file_actions = [
-            (os.POSIX_SPAWN_OPEN, 1, outfile,
-                os.O_WRONLY | os.O_CREAT | os.O_TRUNC,
-                stat.S_IRUSR | stat.S_IWUSR),
-        ]
-        args = self.python_args('-c', script)
-        pid = self.spawn_func(args[0], args, os.environ,
-                              file_actions=file_actions)
-
-        support.wait_process(pid, exitcode=0)
-        with open(outfile, encoding="utf-8") as f:
-            self.assertEqual(f.read(), 'hello')
-
-    def test_close_file(self):
-        closefile = os_helper.TESTFN
-        self.addCleanup(os_helper.unlink, closefile)
-        script = f"""if 1:
-            import os
-            try:
-                os.fstat(0)
-            except OSError as e:
-                with open({closefile!r}, 'w', encoding='utf-8') as closefile:
-                    closefile.write('is closed %d' % e.errno)
-            """
-        args = self.python_args('-c', script)
-        pid = self.spawn_func(args[0], args, os.environ,
-                              file_actions=[(os.POSIX_SPAWN_CLOSE, 0)])
-
-        support.wait_process(pid, exitcode=0)
-        with open(closefile, encoding="utf-8") as f:
-            self.assertEqual(f.read(), 'is closed %d' % errno.EBADF)
-
-    def test_dup2(self):
-        dupfile = os_helper.TESTFN
-        self.addCleanup(os_helper.unlink, dupfile)
-        script = """if 1:
-            import sys
-            sys.stdout.write("hello")
-            """
-        with open(dupfile, "wb") as childfile:
-            file_actions = [
-                (os.POSIX_SPAWN_DUP2, childfile.fileno(), 1),
-            ]
-            args = self.python_args('-c', script)
-            pid = self.spawn_func(args[0], args, os.environ,
-                                  file_actions=file_actions)
-            support.wait_process(pid, exitcode=0)
-        with open(dupfile, encoding="utf-8") as f:
-            self.assertEqual(f.read(), 'hello')
-
-
-@unittest.skipUnless(hasattr(os, 'posix_spawn'), "test needs os.posix_spawn")
-@support.requires_subprocess()
-class TestPosixSpawn(unittest.TestCase, _PosixSpawnMixin):
-    spawn_func = getattr(posix, 'posix_spawn', None)
-
-
-@unittest.skipUnless(hasattr(os, 'posix_spawnp'), "test needs os.posix_spawnp")
-@support.requires_subprocess()
-class TestPosixSpawnP(unittest.TestCase, _PosixSpawnMixin):
-    spawn_func = getattr(posix, 'posix_spawnp', None)
-
-    @os_helper.skip_unless_symlink
-    def test_posix_spawnp(self):
-        # Use a symlink to create a program in its own temporary directory
-        temp_dir = tempfile.mkdtemp()
-        self.addCleanup(os_helper.rmtree, temp_dir)
-
-        program = 'posix_spawnp_test_program.exe'
-        program_fullpath = os.path.join(temp_dir, program)
-        os.symlink(sys.executable, program_fullpath)
-
-        try:
-            path = os.pathsep.join((temp_dir, os.environ['PATH']))
-        except KeyError:
-            path = temp_dir   # PATH is not set
-
-        spawn_args = (program, '-I', '-S', '-c', 'pass')
-        code = textwrap.dedent("""
-            import os
-            from test import support
-
-            args = %a
-            pid = os.posix_spawnp(args[0], args, os.environ)
-
-            support.wait_process(pid, exitcode=0)
-        """ % (spawn_args,))
-
-        # Use a subprocess to test os.posix_spawnp() with a modified PATH
-        # environment variable: posix_spawnp() uses the current environment
-        # to locate the program, not its environment argument.
-        args = ('-c', code)
-        assert_python_ok(*args, PATH=path)
 
 
 @unittest.skipUnless(sys.platform == "darwin", "test weak linking on macOS")
